requirements.txt
Pillow
pytesseract
Flask
kivy
